+++ 10up-experience.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: 10up Experience
  * Description: The 10up Experience plugin configures WordPress to better protect and inform clients, aligned to 10up’s best practices.
- * Version:     1.6.2
+ * Version:     1.7
  * Author:      10up
  * Author URI:  https://10up.com
  * License:     GPLv2 or later
@@ -14,24 +14,33 @@
 
 namespace TenUpExperience;
 
-define( 'TENUP_EXPERIENCE_VERSION', '1.6.2' );
+use Puc_v4_Factory;
 
-require_once __DIR__ . '/includes/utils.php';
-require_once __DIR__ . '/includes/admin.php';
-require_once __DIR__ . '/includes/admin-bar.php';
-require_once __DIR__ . '/includes/admin-pages.php';
-require_once __DIR__ . '/includes/plugins.php';
-require_once __DIR__ . '/includes/rest-api.php';
-require_once __DIR__ . '/includes/gutenberg.php';
-require_once __DIR__ . '/includes/authors.php';
-require_once __DIR__ . '/includes/authentication.php';
-require_once __DIR__ . '/includes/password-protection.php';
-require_once __DIR__ . '/includes/support-monitor.php';
-require_once __DIR__ . '/includes/support-monitor-debug.php';
+define( 'TENUP_EXPERIENCE_VERSION', '1.7' );
+define( 'TENUP_EXPERIENCE_DIR', __DIR__ );
+define( 'TENUP_EXPERIENCE_FILE', __FILE__ );
 
 require_once __DIR__ . '/vendor/yahnis-elsts/plugin-update-checker/plugin-update-checker.php';
 
-$tenup_plugin_updater = \Puc_v4_Factory::buildUpdateChecker(
+require_once __DIR__ . '/includes/utils.php';
+
+spl_autoload_register( function( $class_name ) {
+	$path_parts = explode( '\\', $class_name );
+
+	if ( ! empty( $path_parts ) ) {
+		$package = $path_parts[0];
+
+		unset( $path_parts[0] );
+
+		if ( 'TenUpExperience' === $package ) {
+			require_once __DIR__ . '/includes/classes/' . implode( '/', $path_parts ) . '.php';
+		} elseif ( 'ZxcvbnPhp' === $package ) {
+			require_once __DIR__ . '/vendor/bjeavons/zxcvbn-php/src/' . implode( '/', $path_parts ) . '.php';
+		}
+	}
+} );
+
+$tenup_plugin_updater = Puc_v4_Factory::buildUpdateChecker(
 	'https://github.com/10up/10up-experience/',
 	__FILE__,
 	'10up-experience'
@@ -56,14 +65,13 @@ $network_activated = Utils\is_network_activated( plugin_basename( __FILE__ ) );
 
 define( 'TENUP_EXPERIENCE_IS_NETWORK', (bool) $network_activated );
 
-Admin\setup();
-AdminBar\setup();
-AdminPages\setup();
-Plugins\setup();
-RestAPI\setup();
-Gutenberg\setup();
-Authors\setup();
-Authentication\setup();
-PasswordProtection\setup();
-SupportMonitor\setup();
-
+AdminCustomizations\Customizations::instance()->setup();
+API\API::instance()->setup();
+Authentication\Passwords::instance()->setup();
+Authentication\Usernames::instance()->setup();
+Authors\Authors::instance()->setup();
+Gutenberg\Gutenberg::instance()->setup();
+Plugins\Plugins::instance()->setup();
+PostPasswords\PostPasswords::instance()->setup();
+SupportMonitor\Monitor::instance()->setup();
+SupportMonitor\Debug::instance()->setup();