+++ config/webpack.settings.js
@@ -1,0 +1,68 @@
+// Webpack settings exports.
+module.exports = {
+	entries: {
+		// JS files.
+		passwords: './assets/js/passwords.js',
+
+		// CSS files.
+		admin: './assets/css/admin.css',
+		'passwords-styles': './assets/css/passwords.css',
+		'tenup-pages': './assets/css/tenup-pages.css',
+	},
+	filename: {
+		js: 'js/[name].js',
+		css: 'css/[name].css',
+	},
+	paths: {
+		src: {
+			base: './assets/',
+			css: './assets/css/',
+			js: './assets/js/',
+		},
+		dist: {
+			base: './dist/',
+			clean: ['./images', './css', './js'],
+		},
+	},
+	stats: {
+		// Copied from `'minimal'`.
+		all: false,
+		errors: true,
+		maxModules: 0,
+		modules: true,
+		warnings: true,
+		// Our additional options.
+		assets: true,
+		errorDetails: true,
+		excludeAssets: /\.(jpe?g|png|gif|svg|woff|woff2)$/i,
+		moduleTrace: true,
+		performance: true,
+	},
+	copyWebpackConfig: {
+		from: '**/*.{jpg,jpeg,png,gif,svg,eot,ttf,woff,woff2}',
+		to: '[path][name].[ext]',
+	},
+	ImageminPlugin: {
+		test: /\.(jpe?g|png|gif|svg)$/i,
+	},
+	BrowserSyncConfig: {
+		host: 'localhost',
+		port: 3000,
+		proxy: 'http://10up-experience.test',
+		open: false,
+		files: [
+			'**/*.php',
+			'dist/js/**/*.js',
+			'dist/css/**/*.css',
+			'dist/svg/**/*.svg',
+			'dist/images/**/*.{jpg,jpeg,png,gif}',
+			'dist/fonts/**/*.{eot,ttf,woff,woff2,svg}',
+		],
+	},
+	performance: {
+		maxAssetSize: 100000,
+	},
+	manifestConfig: {
+		basePath: '',
+	},
+};